Do not overwrite existing coordinates if substructure match fails
