+++ packages/Chem/package.json
@@ -5,7 +5,7 @@
   "description": "Cheminformatics support: \n\n* RDKit-based rendering \n* Molecule sketching \n*",
   "sources": [
     "css/chem.css",
-    "src/RDKit_minimal_2021.03_05.js",
+    "src/RDKit_minimal_2021.03_06.js",
     "src/rdkit_worker_proxy.js",
     "src/rdkit_substruct_library.js",
     "src/rdkit_parallel.js",
